# docs/web.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Classification Studio",
    "version": "1.0"
  },
  "paths": {
...
    "/api/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_api_events_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 64,
              "default": "",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/train": {
      "post": {
        "summary": "Train",
        "operationId": "train_api_train_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TrainRequest": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "experiment_name": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Name"
          },
          "epochs": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Epochs",
            "default": 10
          },
          "batch_size": {
            "type": "integer",
            "maximum": 512.0,
            "minimum": 1.0,
            "title": "Batch Size",
            "default": 4
          },
          "learning_rate": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Learning Rate",
            "default": 0.0003
          },
          "num_workers": {
            "type": "integer",
            "maximum": 16.0,
            "minimum": 0.0,
            "title": "Num Workers",
            "default": 0
          },
          "cpu_threads": {
            "type": "integer",
            "maximum": 32.0,
            "minimum": 1.0,
            "title": "Cpu Threads",
            "default": 2
          },
          "device": {
            "type": "string",
            "enum": [
              "global",
              "cpu",
              "gpu",
              "auto"
            ],
            "title": "Device",
            "default": "global"
          },
          "seed": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Seed",
            "default": 42
          },
          "training_mode": {
            "type": "string",
            "enum": [
              "full_finetune",
              "linear_probe"
            ],
            "title": "Training Mode",
            "default": "full_finetune"
          },
          "initialization": {
            "type": "string",
            "enum": [
              "last",
              "official"
            ],
            "title": "Initialization",
            "default": "last"
          },
          "offline": {
            "type": "boolean",
            "title": "Offline",
            "default": false
          },
          "priority": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Priority",
            "default": 0
          },
          "timeout_minutes": {
            "type": "integer",
            "maximum": 10080.0,
            "minimum": 0.0,
            "title": "Timeout Minutes",
            "default": 0
          },
          "gpu_index": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 127.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Index"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "model_id",
          "dataset_id"
        ],
        "title": "TrainRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}